requirements.txt
django>=4.0,<5.0
psycopg2-binary
django-environ
djangorestframework
django-filter
dj-rest-auth
django-cors-headers
django-private-storage
drf-yasg